% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimaond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamond Plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframd df}

\item{vary}{a string name for variable y in the dataframe df}
}
\value{
A plot with diamond markers of \code{vary} versus \code{varx}
}
\description{
This function produces a ggplot plot of \code{vary} versus \code{varx} with diamond markers.
}
\examples{
df.test <- data.frame(x = rnorm(10), y=rnorm(10))

diamond_plot(df.test, 'x', 'y')

}
